GridPage GUI improvements to MidiClock elements

+++ avr/cores/megacommand/Midi/MidiClock.h
@@ -52,7 +52,7 @@ public:
   volatile uint16_t div192th_time;
   volatile uint16_t last_clock16;
 
-  volatile uint16_t bar_counter;
+  volatile uint8_t bar_counter;
   volatile uint8_t beat_counter;
   volatile uint8_t step_counter;
